modify upnp structure
