+++ src/test/scala/bifrost/api/ProgramRPCSpec.scala
@@ -21,6 +21,7 @@ import bifrost.modifier.transaction.bifrostTransaction.Transaction
 import bifrost.network.message._
 import bifrost.network.peer.PeerManagerRef
 import bifrost.network._
+import bifrost.network.upnp.{UPnP, UPnPGateway}
 import bifrost.nodeView.GenericNodeViewHolder.CurrentView
 import bifrost.nodeView.GenericNodeViewHolder.ReceivableMessages.GetDataFromCurrentView
 import bifrost.nodeView.{NodeViewHolderRef, NodeViewModifier}
